requirements.txt
asttokens==2.4.1
batchgenerators==0.23
blobfile==1.2.7
brats==0.0.3
certifi==2021.10.8
charset-normalizer==2.0.8
comm==0.2.2
cycler==0.11.0
dacite==1.8.1
debugpy==1.8.1
decorator==5.1.1
dicom2nifti==2.3.0
docker==7.1.0
docopt==0.6.2
et_xmlfile==2.0.0+computecanada
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.4.2
fonttools==4.28.2
future==0.18.2
h5py==3.8.0
idna==3.3
imageio==2.36.1
imbalanced-learn==0.11.0
imblearn==0.0
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
joblib==1.3.1
jsonpatch==1.32
jsonpointer==2.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.3.2
lightgbm==4.0.0+computecanada
linecache2==1.0.0
llvmlite==0.41.0+computecanada
loguru==0.7.2
markdown-it-py==3.0.0
matplotlib==3.5.0
matplotlib-inline==0.1.7
mdurl==0.1.2
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
nest-asyncio==1.6.0
networkx==2.6.3
nibabel==3.2.1
numba==0.58.0+computecanada
numpy==1.21.4
opencv-python==4.5.5.62
openpyxl==3.1.5+computecanada
packaging==21.3
pandas==1.3.4
parso==0.8.4
patsy==1.0.1+computecanada
pexpect==4.9.0
Pillow==8.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
protobuf==6.30.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodomex==3.12.0
pydicom==2.2.2
Pygments==2.18.0
pykwalify==1.8.0
pynndescent==0.5.13+computecanada
pyparsing==3.0.6
pyradiomics==3.1.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.26.0
rich==13.7.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.18.3
scikit-learn==1.6.0
scipy==1.7.3
seaborn==0.13.2+computecanada
setuptools-scm==6.3.2
shapely @ file:///croot/shapely_1680636282912/work
SimpleITK==2.1.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
stack-data==0.6.3
statsmodels==0.14.0+computecanada
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchfile==0.1.0
torchvision==0.8.2
tornado==6.4
tqdm==4.66.4
traceback2==1.4.0
traitlets==5.14.3
typing_extensions==4.0.1
umap==0.1.1
umap_learn==0.5.7+computecanada
unittest2==1.1.0
urllib3==1.26.7
visdom==0.1.8.9
wcwidth==0.2.13
websocket-client==1.2.3
xgboost==1.7.6
xlrd==2.0.1+computecanada
xmltodict==0.12.0
zipp==3.19.0
